Update README.md
Lines 41 and 66 - `.addTo(map)` is required in real life, is there a reason why this is not in the README?

Because `map` can really mean any variable, I added in the `var map = L.map('map')` lines to indicate where it is coming from and what it is referring to. There may be better ways to do this.

Line 38 (and also elsewhere) - Is it clear what the `[...]` means? Is it necessary? Why is it used in some cases and not others? Could it be clearer as comments so that the code, when copy-pasted, will not break someone's existing code?

Others
- propose `<my api key>` as a replacement placeholder
- use syntax highlighting