"More info"
Changes in line 50. I am not exactly happy with the state of what I added, but I wanted to point out that some of the routing API is really from LRM (e.g. the `waypoints` parameter). So someone hunting in Valhalla documentation for LRM features is going to be very unhappy. That being said I don't know enough yet about what is supposed to be on LRM and what is supposed to be Valhalla to make any clearer distinction than this sentence.